webauthn/getcredential-attachment.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>navigator.credentials.get() authenticator attachment test</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src=helpers.js></script>
<body></body>
<script>
    // usb transport
    virtualAuthenticatorPromiseTest(async function() {
        "use strict";
            const assertion = await navigator.credentials.get({publicKey: {
                    challenge: new Uint8Array(),
                    authenticatorSelection: {
                        authenticatorAttachment: "cross-platform",
                        userVerification: "preferred"
                    },
                    allowCredentials: [{
                        id: (await createCredential()).rawId,
                        type: "public-key",
                    }],
                }});
            assert_equals(assertion.authenticatorAttachment, "cross-platform")
            assert_not_equals(assertion.authenticatorAttachment, null);
            assert_not_equals(assertion.authenticatorAttachment, "platform");
    }, {
        protocol: "ctap2_1",
        transport: "usb"
    }, "navigator.credentials.get() with usb authenticator, attachment as cross-platform");

    // nfc transport without authenticator selection
    virtualAuthenticatorPromiseTest(async function() {
        "use strict";
        const assertion = await navigator.credentials.get({publicKey: {
                challenge: new Uint8Array(),
                authenticatorSelection: {
                    userVerification: "preferred"
                },
                allowCredentials: [{
                    id: (await createCredential()).rawId,
                    type: "public-key",
                }],
            }});
        assert_equals(assertion.authenticatorAttachment, "cross-platform")
        assert_not_equals(assertion.authenticatorAttachment, null);
        assert_not_equals(assertion.authenticatorAttachment, "platform");
    }, {
        protocol: "ctap2_1",
        transport: "nfc"
    }, "navigator.credentials.get() with nfc authenticator, attachment as cross-platform");

    // ble transport without authenticator selection
    virtualAuthenticatorPromiseTest(async function() {
        "use strict";
        const assertion = await navigator.credentials.get({publicKey: {
                challenge: new Uint8Array(),
                authenticatorSelection: {
                    userVerification: "preferred"
                },
                allowCredentials: [{
                    id: (await createCredential()).rawId,
                    type: "public-key",
                }],
            }});
        assert_equals(assertion.authenticatorAttachment, "cross-platform")
        assert_not_equals(assertion.authenticatorAttachment, null);
        assert_not_equals(assertion.authenticatorAttachment, "platform");
    }, {
        protocol: "ctap2_1",
        transport: "ble"
    }, "navigator.credentials.get() with ble authenticator, attachment as cross-platform");

    // internal transport, with authenticator selection preference
    virtualAuthenticatorPromiseTest(async function() {
        "use strict";
            const assertion = await navigator.credentials.get({publicKey: {
                    challenge: new Uint8Array(),
                    authenticatorSelection: {
                        authenticatorAttachment: "platform",
                        userVerification: "preferred"
                    },
                    allowCredentials: [{
                        id: (await createCredential()).rawId,
                        type: "public-key",
                    }],
                }});
            assert_equals(assertion.authenticatorAttachment, "platform")
            assert_not_equals(assertion.authenticatorAttachment, null);
            assert_not_equals(assertion.authenticatorAttachment, "cross-platform");
    }, {
        protocol: "ctap2_1",
        transport: "internal"
    }, "navigator.credentials.get() with internal authenticator, attachment as platform");

    // internal transport, without authenticator selection preference
    virtualAuthenticatorPromiseTest(async function() {
        "use strict";
        const assertion = await navigator.credentials.get({publicKey: {
                challenge: new Uint8Array(),
                authenticatorSelection: {
                    userVerification: "preferred"
                },
                allowCredentials: [{
                    id: (await createCredential()).rawId,
                    type: "public-key",
                }],
            }});
        assert_equals(assertion.authenticatorAttachment, "platform")
        assert_not_equals(assertion.authenticatorAttachment, null);
        assert_not_equals(assertion.authenticatorAttachment, "cross-platform");
    }, {
        protocol: "ctap2_1",
        transport: "internal"
    }, "navigator.credentials.get() with internal authenticator, attachment as platform inferred");
</script>
